perf: improve eslint diagnostic in nvim
As per [1], null-ls's eslint diagnostic spawn multiple eslint
instance when there are changes in normal mode, but it is very
resources intensive, do multiple undo in a row and we see the
effects. Luckily, as in [1] null-ls's author did leave a hint
to use eslint language server provided by lsp-config, and, i
works wonder.

References:
[1] https://github.com/jose-elias-alvarez/null-ls.nvim/issues/596#issuecomment-1024412448
GitOrigin-RevId: cc669e9354bb1c5cbf22b62a5b1a9f8277e1926f