chore(nvim): remove unused keymap
Since I'm using Glove80 with a Engram layout, moving around with $ ^ %
has become so easy, I don't need to use gl, gs, gm from helix anymore.
Esc is also easy, so jk to esc is not needed
Since I'm using wezterm and a window manager, spawning new terminal and
moving them around is so much easier, so keymap used to moving around
Neovim built-in terminal is not needed anymore.
Wezterm also has support for image protocol, so imv on Linux is not
needed
Markdown previewing keymap might be worth keeping, but a markdown is
easy to view and edit raw, so I don't see myself using this feature much