feat(nvim): try oil.nvim as replacement for neotree
Neotree has the following problems:
- On files with really long name or deep file hierarchy, it will
truncate the file name with no option to view the entire file name
- It open and close folder in the tree view at random
- Can't copy file name
- Creating new files is awkward, since there is limited options in the
new file creating UI to edit text
Beside, most of the times, when I need to use neotree, I just need to
browse relatively close in the hierarchical, and oil fits that
perfectly.
If I need to browse far, telescope is enough.